requirements.txt
fastapi[standard]

python-jose 

python-dotenv

pydantic